refactor(in-app): refactor api with accept-language header and extract to /api
