+++ in-app/v1/src/api/ticket.ts
@@ -1,0 +1,31 @@
+import { http } from '@/leancloud';
+import { Ticket } from '@/types';
+import { UseQueryOptions, useQuery } from 'react-query';
+
+async function fetchTicket(id: string): Promise<Ticket> {
+  const { data } = await http.get('/api/2/tickets/' + id);
+  return data;
+}
+
+export function useTicket(id: string, options?: UseQueryOptions<Ticket>) {
+  return useQuery({
+    queryKey: ['ticket', id],
+    queryFn: () => fetchTicket(id),
+    ...options,
+  });
+}
+async function fetchUnread(categoryId?: string) {
+  const { data } = await http.get<boolean>(`/api/2/unread`, {
+    params: {
+      product: categoryId,
+    },
+  });
+  return data;
+}
+
+export function useHasUnreadTickets(categoryId?: string) {
+  return useQuery({
+    queryKey: ['unread', categoryId],
+    queryFn: () => fetchUnread(categoryId),
+  });
+}
